+++ HKMP/Networking/Server/NetServerClient.cs
@@ -3,15 +3,15 @@ using System.Net;
 using System.Net.Sockets;
 
 namespace HKMP.Networking.Server {
-    public delegate void OnReceive(int id, List<Packet.Packet> packets);
+    public delegate void OnReceive(ushort id, List<Packet.Packet> packets);
     /**
      * A client managed by the server.
      * This is only used for communication from server to client.
      */
     public class NetServerClient {
-        private static int _lastId = 0;
+        private static ushort _lastId = 0;
 
-        private readonly int _id;
+        private readonly ushort _id;
         private readonly TcpNetClient _tcpNetClient;
 
         private readonly IPEndPoint _endPoint;
@@ -21,7 +21,7 @@ namespace HKMP.Networking.Server {
         public NetServerClient(TcpClient tcpClient) : this(_lastId++, tcpClient) {
         }
 
-        public NetServerClient(int id, TcpClient tcpClient) {
+        public NetServerClient(ushort id, TcpClient tcpClient) {
             _id = id;
 
             // Create UDP endpoint with TCP address and UDP port
@@ -64,7 +64,7 @@ namespace HKMP.Networking.Server {
             return _endPoint.Address.Equals(endPoint.Address);
         }
 
-        public int GetId() {
+        public ushort GetId() {
             return _id;
         }
